% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del_data.R
\name{del_data}
\alias{del_data}
\title{Create missing data}
\usage{
del_data(df, mType = 1, mPer = 0.5, propDel = 0.1)
}
\arguments{
\item{df}{Data frame. Data frame of longitudinal
data.}

\item{mType}{Numeric. Type of missingness. 1 = random
missingness; 2 = non-random missingness. Default is 1.}

\item{mPer}{Numeric. Proportion of missing data. Range
(0, 1). If mType = 1, then randomly deletes mPer\emph{100\%
of rows from who data frame. If mType = 2, deletes
mPer}100\% rows from the groups selected to
have missingness. Default is 0.5.}

\item{propDel}{Numeric. Range (0,1). Only applies
if mType=2. Denotes proportion of groups from which
data is to be missing. Default is 0.1.}
}
\value{
Data frame with rows deleted
}
\description{
Deletes records from longitudinal data frame
in a random or non-random manner. For random deletion,
function deletes a designated percent of rows at random.
If deletion is random, the function specifies a subset
of groups to have missing data, and then deletes a
designated percent of rows from only those groups.
(Note: In longitudinal data, a "group" is an individual
with repeated measurements). Function designed so that
at least one record (row) will be deleted.
}
\examples{
del_data(df, mType=1, mPer=0.05)
del_data(df, mType=2, mPer=0.1, propDel=0.5)
}
