% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_time.R
\name{make_time}
\alias{make_time}
\title{Define timepoints}
\usage{
make_time(clus, waves, tType = 1, incr = 1, mint = 1, maxt = 10)
}
\arguments{
\item{clus}{Numeric. Number of samples/clusters.}

\item{waves}{Numeric. Number of timepoints}

\item{tType}{Numeric. Type of interval spacing. If tType=1,
equal intervals. If tType=2, increasing intervals.
If tType=3, random intervals. Default is 1.}

\item{incr}{Numeric. The increment of time spacing when
tType=1 or tType=3. Default is 1.}

\item{mint}{Numeric. Lower limit of time, if tType=3.
Must be greater than 0, but less than maxt. Default is 1.}

\item{maxt}{Numeric. Upper limit of time, if tType=3.
Must be greater than mint. Default is 10.}
}
\value{
Vector of timepoints, grouped by cluster.
}
\description{
Creates the timepoints (e.g. waves, measurement occasions)
for simulating longitudinal data.
}
\details{
There are three options for time interval spacing.
\itemize{
\item \emph{Equal} - Time intervals are equal and even.
All individuals have the same timepoints.
\item \emph{Increasing} - Time intervals increase with
each wave. All individuals have the same timepoints.
\item \emph{Random} - Time intervals are random, selected
from a uniform distribution. Each individual has
different timepoints
}
}
\examples{
make_time(clus=10, waves=3, tType=1)
make_time(clus=10, waves=3, tType=2, incr=5)
make_time(clus=10, waves=3, tType=3)
}
